adding authentication

+++ graphql/resolvers/booking.js
@@ -8,19 +8,25 @@ import Booking from '../../models/booking';
 import { transformEvent, transformBooking } from './merge';
 
 export default {
-  bookings: async () => {
+  bookings: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const bookings = await Booking.find();
       return bookings.map((booking) => transformBooking(booking));
     } catch (err) {
       throw err;
     }
   },
-  bookEvent: async (args) => {
+  bookEvent: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const fetchedEvent = await Event.findOne({ _id: args.eventId });
       const booking = new Booking({
-        user: '5f71fd39b618544280412b78',
+        user: req.userId,
         event: fetchedEvent,
       });
       const result = await booking.save();
@@ -29,8 +35,11 @@ export default {
       throw err;
     }
   },
-  cancelBooking: async (args) => {
+  cancelBooking: async (args, req) => {
     try {
+      if (!req.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       const booking = await Booking.findById(args.bookingId).populate('event');
       const event = transformEvent(booking.event);
       await Booking.deleteOne({ _id: args.bookingId });
